Safari-611.1.10/LayoutTests/http/tests/inspector/network/resource-response-inspector-override-expected.txt

Test for `Resource.ResponseSource.InspectorOverride`.


== Running test suite: Resource.ResponseSource.InspectorOverride
-- Running test setup.
-- Running test case: Resource.ResponseSource.InspectorOverride
PASS: Resource should be created.
PASS: Resource should receive a Response.
PASS: statusCode should be 987
PASS: responseSource should be Symbol(inspector-override)
